split Credential into GitHubServerConfig + AccessTolenCredential
