+++ src/main/resources/com/cloudbees/jenkins/Credential/config.jelly
@@ -1,12 +1,8 @@
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
+<j:jelly xmlns:j="jelly:core" xmlns:c="/lib/credentials" xmlns:f="/lib/form">
   <f:entry title="${%API URL}" field="apiUrl">
     <f:textbox/>
   </f:entry>
-  <f:entry title="${%Username}" field="username">
-    <f:textbox/>
+  <f:entry title="${%Credentials}" field="credentialId">
+    <c:select/>
   </f:entry>
-  <f:entry title="${%OAuth token}" field="oauthAccessToken">
-    <f:textbox/>
-  </f:entry>
-  <f:validateButton title="${%Test Credential}" with="apiUrl,username,password,oauthAccessToken" method="validate"/>
 </j:jelly>
